moved file to inside of components
